+++ src/routes/income.routes.js
@@ -2,7 +2,7 @@ const express = require('express')
 
 const router = express.Router();
 
-const incomeController = require('../controllers/income.controller');
+const incomeController = require('../../controllers/income.controller');
 
 router.get('/total-earned', incomeController.getTotalEarnedForAYear); // query params: year
